camunits-extra debian packaging work

+++ debian/camunits-extra/from-svn.py
@@ -11,41 +11,44 @@ from xml.dom import minidom
 def do_or_die(cmd):
     print "Running [%s]" % cmd
     if 0 != os.system(cmd):
-        raise RuntimeError ("damn.")
+        raise RuntimeError("damn.")
 
-if len (sys.argv) > 1:
-    print ("usage: %s" % (os.path.basename(sys.argv[0])))
+if len(sys.argv) > 1:
+    print("usage: %s" % (os.path.basename(sys.argv[0])))
     sys.exit(1)
+
+dont_use_cache = False
 
 package_name = "camunits-extra"
 svn_url = "https://camunits.googlecode.com/svn/plugins/camunits-extra"
 
-do_or_die ("svn update")
+do_or_die("svn update")
 status, xml_str = commands.getstatusoutput("svn info --xml %s" % svn_url)
 
-info = minidom.parseString (xml_str)
-entry_node = info.documentElement.getElementsByTagName ("entry")[0]
-svn_revision = int (entry_node.getAttribute("revision"))
+info = minidom.parseString(xml_str)
+entry_node = info.documentElement.getElementsByTagName("entry")[0]
+svn_revision = int(entry_node.getAttribute("revision"))
 
 # start the process...
 print "Cleaning out build-area"
-do_or_die ("rm -rf build-area/*")
+do_or_die("rm -rf build-area/*")
 
-print ("Reverting trunk")
-do_or_die ("svn revert --recursive trunk")
+print("Reverting changelog")
+do_or_die("svn revert trunk/debian/changelog")
 os.chdir("trunk")
 
 export_dir = "%s.svn%d" % (package_name, svn_revision)
-if os.path.exists (export_dir):
+if os.path.exists(export_dir) and dont_use_cache:
     print "Clearing out %s" % export_dir
-    do_or_die ("rm -rf %s" % export_dir)
+    do_or_die("rm -rf %s" % export_dir)
 
-print ("Checking out a copy of source tree")
-do_or_die ("svn export -r %d %s %s" % (svn_revision, svn_url, export_dir))
+if not os.path.exists(export_dir):
+    print("Checking out a copy of source tree")
+    do_or_die("svn export -r %d %s %s" % (svn_revision, svn_url, export_dir))
 
-os.chdir (export_dir)
+os.chdir(export_dir)
 
-print ("Extracting version number")
+print("Extracting version number")
 regex = re.compile("AM_INIT_AUTOMAKE\s*\(\s*%s\s*,\s*([^\s)]+)\s*\)" % \
         package_name)
 release_version = None
@@ -59,24 +62,29 @@ if not release_version:
     sys.exit(1)
 print "release version: %s" % release_version
 
-print ("Building orig.tar.gz")
-do_or_die ("autoreconf -i")
-do_or_die ("./configure")
-do_or_die ("make")
-do_or_die ("make distcheck")
-os.chdir ("..")
-orig_tarball = "../tarballs/%s_%s.svn%d.orig.tar.gz" % \
+print("Building orig.tar.gz")
+orig_tarball = "../../tarballs/%s_%s.svn%d.orig.tar.gz" % \
         (package_name, release_version, svn_revision) 
-if os.path.exists (orig_tarball):
-    os.unlink (orig_tarball)
-shutil.copyfile ("%s/%s-%s.tar.gz" % (export_dir, package_name, 
-    release_version), orig_tarball)
+if os.path.exists(orig_tarball) and dont_use_cache:
+    os.unlink(orig_tarball)
 
-print ("Updating debian/changelog")
-do_or_die ("dch --newversion %s.svn%d-1 \"New subversion build\"" % \
+if not os.path.exists(orig_tarball):
+    print os.listdir("../")
+    do_or_die("autoreconf -i")
+    do_or_die("./configure")
+    do_or_die("make")
+    do_or_die("make distcheck")
+
+    shutil.copyfile("%s/%s-%s.tar.gz" % (export_dir, package_name, 
+        release_version), orig_tarball)
+
+os.chdir("..")
+
+print("Updating debian/changelog")
+do_or_die("dch --newversion %s.svn%d-1 \"New subversion build\"" % \
         (release_version, svn_revision))
 
 print "Building..."
-do_or_die ("svn-buildpackage -us -uc -rfakeroot --svn-ignore")
+do_or_die("svn-buildpackage -us -uc -rfakeroot --svn-ignore")
 
-os.system ("svn revert debian/changelog")
+os.system("svn revert debian/changelog")
